src/langgraph-test/pyproject.toml
[project]
name = "langgraph-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3==1.40.69",
    "types-boto3[full]==1.40.69",
    "python-dotenv",
    "langgraph",
    "langchain-aws",
    "chromadb",
    "llama_index"
]
